fix RIDER-35494, remove hack for cancelling tests, move file to corresponding place
