+++ resharper/resharper-unity/src/Rider/UnitTesting/RunViaUnityEditorStrategy.cs
@@ -275,13 +275,8 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.UnitTesting
                                     SubscribeResults(run, lt, tcs, launch);
                                 });
 
-                                myUnityProcessId.When(taskLifetime, (int?)null, _ => tcs.TrySetException(new Exception("Unity Editor has been closed."))); 
+                                myUnityProcessId.When(taskLifetime, (int?)null, _ => tcs.TrySetException(new Exception("Unity Editor has been closed.")));
                                 
-                                // KS: IMHO this should be removed when Abort action will be supported in unity.rider package for Unity 2019.2+
-                                // Otherwise this allows several parallel test runs in Unity Editor, which is not supported on the plugin side, and can
-                                // easily damage per-test coverage data.
-                                taskLifetime.OnTermination(cancellationTs.Token.Register(() => tcs.TrySetCanceled()));
-
                                 var rdTask = myEditorProtocol.UnityModel.Value.RunUnitTestLaunch.Start(Unit.Instance);
                                 rdTask?.Result.Advise(taskLifetime, res =>
                                 {
@@ -432,7 +427,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Rider.UnitTesting
 
                 myConnectionTracker.State.When(waitingLifetimeDef.Lifetime, UnityEditorState.Idle, _ => tcs.TrySetResult(Unit.Instance));
                 myUnityProcessId.When(waitingLifetimeDef.Lifetime, (int?)null, _ => tcs.TrySetException(new Exception("Unity Editor has been closed.")));
-                waitingLifetimeDef.Lifetime.OnTermination(() => cancellationToken.Register(() => tcs.TrySetCanceled()));
+                waitingLifetimeDef.Lifetime.TryOnTermination(() => cancellationToken.Register(() => tcs.TrySetCanceled()));
             });
 
             return tcs.Task;
